requirements.txt
Flask==3.1.0
Flask-Migrate==4.0.7
Werkzeug==3.1.3
PyJWT==2.10.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
pytest==8.3.4